Added Level Creation

+++ mario/NewSprite.h
@@ -6,7 +6,7 @@ public:
 	int texture_id;
 	int texture_id2=-1;
 	int death=-1;
-	float velx=0, vely=0;
+	int velx=0, vely=0;
 	int xl=0, xr=0, yt=0, yb=0;
 
 	sf::Sprite* sprite;
